// src/types/dispatch.interface.ts
import { RowDataPacket } from 'mysql2';
import { Customer } from './document.interface';

export interface DispatchRow extends RowDataPacket {
    id: number;
    website_id: number | null;
    company: string | null;
    user_id: number;
    external_id: string;
    establishment_id: number;
    inventory_reference_id: number | null;
    establishment: string;
    soap_type_id: string;
    state_type_id: string;
    ubl_version: string;
    document_type_id: string;
    reference_sale_note_id: number | null;
    reference_document_id: number | null;
    reference_dispatch_order_id: number | null;
    reference_quotation_id: number | null;
    reference_order_form_id: number | null;
    reference_order_note_id: number | null;
    series: string;
    number: number;
    date_of_issue: Date;
    time_of_issue: string;
    customer_id: number;
    customer: Customer;
    observations: string;
    transport_mode_type_id: string;
    transfer_reason_type_id: string;
    transfer_reason_description: string | null;
    date_of_shipping: Date;
    transshipment_indicator: number;
    port_code: string | null;
    unit_type_id: string;
    total_weight: string;
    packages_number: number;
    container_number: string | null;
    related: any;
    origin: string;
    delivery: string;
    dispatcher_id: number | null;
    dispatcher: any;
    driver_id: number;
    secondary_driver_id: number | null;
    driver: string;
    order_form_external: string | null;
    license_plate: string | null;
    sender_id: number | null;
    sender_data: any;
    sender_address_id: number | null;
    sender_address_data: any;
    receiver_id: number | null;
    receiver_data: any;
    receiver_address_id: number | null;
    receiver_address_data: any;
    transport_id: number;
    secondary_transport_id: number | null;
    transport_data: string;
    secondary_license_plates: string;
    legends: string;
    optional: any;
    filename: string;
    hash: string | null;
    soap_shipping_response: string | null;
    send_to_pse: number;
    response_signature_pse: string | null;
    response_send_cdr_pse: string | null;
    has_xml: number;
    has_pdf: number;
    has_cdr: number;
    document_id: number | null;
    ticket: string;
    message_ticket: string;
    reception_date: string;
    qr_url: string;
    created_at: Date;
    updated_at: Date;
    data_affected_document: any;
    terms_condition: any;
    additional_data: any;
    origin_address_id: number;
    delivery_address_id: number;
    tracto_carreta: any;
    purchase_order: any;
    alter_company: any;
    secondary_driver: any;
    secondary_transport_data: any;
    flete_payment_number: any;
    flete_payment_name: any;
    flete_payer_id: any;
    supplier_id: number;
    plate_number: string | null;
    is_transport_category_m1l: number;
    related_document: any;
    subcontract_company_id: number | null;
    auditor_state: number;
    attempt_pse: number;
    migrated: number;
    migrated_file: number;
    migrated_at: Date | null;
    migrated_updated_at: Date | null;
    migrated_status_code: string | null;
    migrated_document_id: string | null;
}